Moved sky_line_fwhm to spec_functions
